+++ packages/terra-compact-interactive-list/src/utils/utils.js
@@ -1,4 +1,4 @@
-import { widthUnitTypes, WARNINGS } from './constants';
+import { widthUnitTypes, DefaultListValues, WARNINGS } from './constants';
 
 /**
  * A method that splits the css width, mimWidth or Maxwidth string into number value and string type.
@@ -29,7 +29,7 @@ export const getValueUnitTypePair = (valueString) => {
 export const checkIfColumnIsResponsive = (flexGrow, width) => flexGrow || !(width && typeof width === 'number');
 
 // A row is considered being responsive if it has at least one column that can grow or shrink (a flex column).
-const checkIfAnyColumnIsResponsive = (total, column) => total || checkIfColumnIsResponsive(column.flexGrow, column.width);
+const checkIfAnyColumnIsResponsive = (total, column) => total || checkIfColumnIsResponsive(column?.flexGrow, column?.width);
 export const checkIfRowHasResponsiveColumns = (columns) => columns.reduce(checkIfAnyColumnIsResponsive, false);
 
 /**
@@ -41,7 +41,7 @@ export const converseColumnTypes = (columns, defaultType) => {
   // get unitType and check it's consistant across columns
 
   let i = 0;
-  while (!unitType && i < columns.length) {
+  while (!unitType && i < columns?.length) {
     const widthProp = columns[i].width || columns[i].minimumWidth || columns[i].maximumWidth;
     if (widthProp) {
       unitType = getValueUnitTypePair(widthProp)?.unitType;
@@ -57,21 +57,30 @@ export const converseColumnTypes = (columns, defaultType) => {
     const width = getValueUnitTypePair(column.width);
     const minimumWidth = getValueUnitTypePair(column.minimumWidth);
     const maximumWidth = getValueUnitTypePair(column.maximumWidth);
-    if (
-      (width && width?.unitType !== unitType)
-      || (minimumWidth && maximumWidth?.unitType !== unitType)
-      || (minimumWidth && minimumWidth?.unitType !== unitType)) {
-      // eslint-disable-next-line no-console
-      console.warn(WARNINGS.COLUMN_WIDTH_INCONSISTENT_TYPE);
-    }
     const newColumn = { ...column };
+    if (width && width?.unitType !== unitType) {
+      delete newColumn.width;
+      // eslint-disable-next-line no-console
+      console.warn(`${WARNINGS.COLUMN_WIDTH_INCONSISTENT_TYPE} ${column.id} column's ${'width'} property unit type is ${width?.unitType}, while it should be ${unitType}. It will be disregarded, which will lead to ${column.id} becoming a responsive column.`);
+    }
+    if (maximumWidth && maximumWidth?.unitType !== unitType) {
+      delete newColumn.maximumWidth;
+      // eslint-disable-next-line no-console
+      console.warn(`${WARNINGS.COLUMN_WIDTH_INCONSISTENT_TYPE} ${column.id} column's ${'maximumWidth'} property unit type is ${maximumWidth?.unitType} instead of ${unitType}. It will be disregarded.`);
+    }
+    if (minimumWidth && minimumWidth?.unitType !== unitType) {
+      delete newColumn.minimumWidth;
+      // eslint-disable-next-line no-console
+      console.warn(`${WARNINGS.COLUMN_WIDTH_INCONSISTENT_TYPE} ${column.id} column's ${'minimumWidth'} property unit type is ${minimumWidth?.unitType} instead of ${unitType}. It will be disregarded, the default value will be used instead.`);
+    }
+
     if (width?.unitType === unitType) {
       newColumn.width = width?.value;
     }
     if (maximumWidth?.unitType === unitType) {
       newColumn.maximumWidth = maximumWidth?.value;
     }
-    if (minimumWidth) {
+    if (minimumWidth?.unitType === unitType) {
       newColumn.minimumWidth = minimumWidth?.value;
     }
     return newColumn;
@@ -88,10 +97,10 @@ export const getRowMaximumWidth = (columns, columnMaximumWidth) => {
   let flexColumnWasFound = false;
   let maxWidth = 0;
   for (let i = 0; i < columns.length; i += 1) {
-    if (checkIfColumnIsResponsive(columns[i].flexGrow, columns[i].width)) {
+    if (checkIfColumnIsResponsive(columns[i]?.flexGrow, columns[i]?.width)) {
       flexColumnWasFound = true;
-      if (columns[i].maximumWidth || columnMaximumWidth) {
-        const colMaxWidth = columns[i].maximumWidth || columnMaximumWidth;
+      if (columns[i]?.maximumWidth || columnMaximumWidth) {
+        const colMaxWidth = columns[i]?.maximumWidth || columnMaximumWidth;
         maxWidth += colMaxWidth;
       } else { return null; }
     } else if (columns[i].width) {
@@ -110,14 +119,38 @@ export const getRowMinimumWidth = (columns, columnMinimumWidth) => {
   let flexColumnWasFound = false;
   let minWidth = 0;
   for (let i = 0; i < columns.length; i += 1) {
-    if (checkIfColumnIsResponsive(columns[i].flexGrow, columns[i].width)) {
+    if (checkIfColumnIsResponsive(columns[i]?.flexGrow, columns[i]?.width)) {
       flexColumnWasFound = true;
-      if (columns[i].minimumWidth || columnMinimumWidth) {
-        minWidth += columns[i].minimumWidth || columnMinimumWidth;
+      if (columns[i]?.minimumWidth || columnMinimumWidth) {
+        minWidth += columns[i]?.minimumWidth || columnMinimumWidth;
       } else { return null; }
     } else if (columns[i].width) {
       minWidth += columns[i].width;
     } else { return null; }
   }
   return flexColumnWasFound ? minWidth : null;
+};
+
+// ensures that columnMinimumWidth is width unit type consistent
+export const getColumnMinWidth = (columnMinimumWidth, widthUnit) => {
+  let columnMinWidth = getValueUnitTypePair(columnMinimumWidth);
+  if (!columnMinWidth) {
+    columnMinWidth = getValueUnitTypePair(DefaultListValues.minimumWidth[widthUnit]);
+  } else if (columnMinWidth.unitType !== widthUnit) {
+    // eslint-disable-next-line no-console
+    console.warn(WARNINGS.COLUMN_MIN_WIDTH_UNIT_TYPE);
+    columnMinWidth = getValueUnitTypePair(DefaultListValues.minimumWidth[widthUnit]);
+  }
+  return columnMinWidth;
+};
+
+// ensures that columnMinimumWidth is width unit type consistent
+export const getColumnMaxWidth = (columnMaximumWidth, widthUnit) => {
+  let columnMaxWidth = getValueUnitTypePair(columnMaximumWidth);
+  if (columnMaxWidth && columnMaxWidth.unitType !== widthUnit) {
+    // eslint-disable-next-line no-console
+    console.warn(WARNINGS.COLUMN_MAX_WIDTH_UNIT_TYPE);
+    columnMaxWidth = null;
+  }
+  return columnMaxWidth;
 };