+++ packages/terra-embedded-content-consumer/tests/wdio/embedded-content-consumer-spec.js
@@ -12,7 +12,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('Provider triggers EventA message', () => {
-      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-event-provider"]');
+      const myFrame = $('iframe[id="custom-event-consumer-frame"]');
       browser.switchToFrame(myFrame);
 
       $('#EventA').click();
@@ -30,7 +30,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('Provider triggers EventA message', () => {
-      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[id="custom-events-consumer-frame"]');
       browser.switchToFrame(myFrame);
 
       $('#EventA').click();
@@ -40,7 +40,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('successfully replied with EventA message', () => {
-      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[id="custom-events-consumer-frame"]');
       browser.switchToFrame(myFrame);
 
       expect($('#embedded-content-consumer-reply')).toHaveTextContaining('eventA');
@@ -48,7 +48,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('Provider triggers EventB message', () => {
-      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[id="custom-events-consumer-frame"]');
       browser.switchToFrame(myFrame);
 
       $('#EventB').click();
@@ -58,7 +58,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('successfully replied with EventB message', () => {
-      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[id="custom-events-consumer-frame"]');
       browser.switchToFrame(myFrame);
 
       expect($('#embedded-content-consumer-reply')).toHaveTextContaining('eventB');
@@ -69,9 +69,9 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     it('has mounted, launched, and authorized elements', () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/data-status-consumer');
       const timeout = browser.options.waitforTimeout + 5000;
-      $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/data-status-provider"]').waitForDisplayed({ timeout });
+      $('iframe[id="data-embedded-consumer-data-status"]').waitForDisplayed({ timeout });
 
-      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/data-status-provider"]');
+      const myFrame = $('iframe[id="data-embedded-consumer-data-status"]');
       browser.switchToFrame(myFrame);
 
       expect($('#Mounted').isExisting());
@@ -103,6 +103,82 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
       browser.switchToParentFrame();
 
       Terra.validates.element('Scroll content into view');
+    });
+  });
+
+  describe('shows visual focus indicator when all criterias are met', () => {
+    it('on the frame when clicked', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/basic-consumer-with-scrolling');
+      $('iframe[id="basic-consumer-with-scrolling"]').waitForDisplayed();
+
+      const myFrame = $('iframe[id="basic-consumer-with-scrolling"]');
+      browser.switchToFrame(myFrame);
+
+      const p = $('<p>');
+      p.scrollIntoView();
+      p.click();
+
+      browser.switchToParentFrame();
+
+      Terra.validates.element('focus indicator on the frame when clicked', { selector: '#site' });
+    });
+
+    it('on the frame when tabbing through and focus is on the frame', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/basic-consumer-with-scrolling');
+      $('iframe[id="basic-consumer-with-scrolling"]').waitForDisplayed();
+      browser.keys('Tab');
+
+      Terra.validates.element('focus indicator on the frame when tabbing through', { selector: '#site' });
+    });
+
+    it('on the frame when tabbing through for inline html content inside of `srcdoc`', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/basic-consumer-with-srcdoc');
+      $('iframe[id="iframe-with-srcdoc"]').waitForDisplayed();
+
+      const myFrame = $('iframe[id="iframe-with-srcdoc"]');
+      browser.switchToFrame(myFrame);
+
+      const p = $('<p>');
+      p.scrollIntoView();
+
+      browser.switchToParentFrame();
+      browser.keys('Tab');
+
+      Terra.validates.element('focus indicator on the frame using inline html with srcdoc', { selector: '#site' });
+    });
+  });
+
+  describe('No visual focus indicator when', () => {
+    it('there is any interactable element in the content', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/custom-events-consumer');
+      $('iframe[id="custom-events-consumer-frame"]').waitForDisplayed();
+
+      const myFrame = $('iframe[id="custom-events-consumer-frame"]');
+      browser.switchToFrame(myFrame);
+
+      const button = $('<button>');
+      button.scrollIntoView();
+
+      browser.switchToParentFrame();
+      browser.keys('Tab');
+
+      Terra.validates.element('no focus indicator on the frame when there is any interactable element', { selector: '#site' });
+    });
+
+    it('scrolling is not enabled on the frame', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/basic-consumer-with-no-scrolling');
+
+      browser.keys('Tab');
+
+      Terra.validates.element('no focus indicator on the frame when scrolling is not enabled', { selector: '#site' });
+    });
+
+    it('the content is not larger than the view port', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/basic-consumer');
+
+      browser.keys('Tab');
+
+      Terra.validates.element('no focus indicator on the frame when content height is not larger than view port', { selector: '#site' });
     });
   });
 });