+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.6.0 - (December 5, 2023)
+
+* Changed
+  * Removed console warning message when no pinned columns exist.
+
+## 1.5.0 - (December 1, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.4.0 - (November 23, 2023)
 
 * Added