+++ packages/terra-menu/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added
   * New stylings for the focus indicator.
+* Fixed
+  * Fixed the issue with `contentWidth` as auto value.
 
 ## 6.81.0 - (November 23, 2023)